Module defindex_strategy_core::event

This module defines the events used in the strategy. These events must be emitted by the strategy implementation to ensure compliance with the expected event interface.

Structs

DepositEvent

HarvestEvent

InitializedEvent

WithdrawEvent

Statics

__SPEC_XDR_TYPE_DEPOSITEVENT

__SPEC_XDR_TYPE_HARVESTEVENT

__SPEC_XDR_TYPE_INITIALIZEDEVENT

__SPEC_XDR_TYPE_WITHDRAWEVENT

Functions

emit_deposit

Publishes an DepositEvent to the event stream.

emit_harvest

Publishes an HarvestEvent to the event stream.

emit_initialize

Publishes an InitializedEvent to the event stream.

emit_withdraw

Publishes an WithdrawEvent to the event stream.