Struct defindex_strategy_core::DeFindexStrategyClient

pub struct DeFindexStrategyClient<'a> {
    pub env: Env,
    pub address: Address,
    /* private fields */
}

DeFindexStrategyClient is a client for calling the contract defined in “DeFindexStrategyTrait”.

Fields
env: Envaddress: Address
Implementations

impl<'a> DeFindexStrategyClient<'a>

pub fn new(env: &Env, address: &Address) -> Self

impl<'a> DeFindexStrategyClient<'a>

pub fn initialize(&self, asset: &Address, init_args: &Vec<Val>)

Initializes the strategy with the given parameters.

This function is called to initialize the strategy with the required parameters. It sets up the strategy’s underlying asset and any additional parameters required for the strategy.

Parameters

env: The execution environment.

asset: The address of the underlying asset that this strategy will manage.

init_args: A vector of additional arguments, which may vary depending on the strategy’s requirements (e.g., configurations, addresses).

Returns

Ok(()) if the strategy is successfully initialized.

Err(StrategyError) if an error occurs during initialization.

pub fn try_initialize( &self, asset: &Address, init_args: &Vec<Val> ) -> Result<Result<(), <() as TryFromVal<Env, Val>>::Error>, Result<StrategyError, InvokeError>>

Initializes the strategy with the given parameters.

This function is called to initialize the strategy with the required parameters. It sets up the strategy’s underlying asset and any additional parameters required for the strategy.

Parameters

env: The execution environment.

asset: The address of the underlying asset that this strategy will manage.

init_args: A vector of additional arguments, which may vary depending on the strategy’s requirements (e.g., configurations, addresses).

Returns

Ok(()) if the strategy is successfully initialized.

Err(StrategyError) if an error occurs during initialization.

pub fn asset(&self) -> Address

Returns the underlying asset address managed by the strategy.

This function provides the address of the asset that this strategy is managing. It is used by the DeFindex vault to determine which asset is associated with this strategy.

Parameters

env: The execution environment.

Returns

Ok(Address) containing the asset address.

Err(StrategyError) if the asset address cannot be retrieved.

pub fn try_asset( &self ) -> Result<Result<Address, <Address as TryFromVal<Env, Val>>::Error>, Result<StrategyError, InvokeError>>

Returns the underlying asset address managed by the strategy.

This function provides the address of the asset that this strategy is managing. It is used by the DeFindex vault to determine which asset is associated with this strategy.

Parameters

env: The execution environment.

Returns

Ok(Address) containing the asset address.

Err(StrategyError) if the asset address cannot be retrieved.

pub fn deposit(&self, amount: &i128, from: &Address)

Deposits the specified amount of the asset into the strategy.

This function allows the DeFindex vault to deposit assets into the strategy.

Parameters

env: The execution environment.

amount: The amount of the underlying asset to deposit into the strategy.

from: The address from which the asset will be transferred (usually the DeFindex vault).

Returns

Ok(()) if the deposit is successful.

Err(StrategyError) if the deposit fails for any reason.

pub fn try_deposit( &self, amount: &i128, from: &Address ) -> Result<Result<(), <() as TryFromVal<Env, Val>>::Error>, Result<StrategyError, InvokeError>>

Deposits the specified amount of the asset into the strategy.

This function allows the DeFindex vault to deposit assets into the strategy.

Parameters

env: The execution environment.

amount: The amount of the underlying asset to deposit into the strategy.

from: The address from which the asset will be transferred (usually the DeFindex vault).

Returns

Ok(()) if the deposit is successful.

Err(StrategyError) if the deposit fails for any reason.

pub fn harvest(&self, from: &Address)

Harvests yields generated by the strategy.

This function triggers any necessary actions required to generate or claim yields from the underlying DeFi protocol or strategy. It can be used to reinvest rewards, claim interest, or perform other actions that maximize returns.

Parameters

env: The execution environment.

from: The address that is triggering the harvest (usually the DeFindex vault).

Returns

Ok(()) if the harvest is successful.

Err(StrategyError) if the harvest fails.

pub fn try_harvest( &self, from: &Address ) -> Result<Result<(), <() as TryFromVal<Env, Val>>::Error>, Result<StrategyError, InvokeError>>

Harvests yields generated by the strategy.

This function triggers any necessary actions required to generate or claim yields from the underlying DeFi protocol or strategy. It can be used to reinvest rewards, claim interest, or perform other actions that maximize returns.

Parameters

env: The execution environment.

from: The address that is triggering the harvest (usually the DeFindex vault).

Returns

Ok(()) if the harvest is successful.

Err(StrategyError) if the harvest fails.

pub fn balance(&self, from: &Address) -> i128

Returns the current balance of the underlying asset in the strategy.

This function provides the current balance of the underlying asset that is managed by the strategy on behalf of the specified address. It is used to track the assets held in the strategy.

Parameters

env: The execution environment.

from: The address for which to retrieve the balance (usually the DeFindex vault).

Returns

Ok(i128) containing the balance of the underlying asset.

Err(StrategyError) if the balance cannot be retrieved.

pub fn try_balance( &self, from: &Address ) -> Result<Result<i128, <i128 as TryFromVal<Env, Val>>::Error>, Result<StrategyError, InvokeError>>

Returns the current balance of the underlying asset in the strategy.

This function provides the current balance of the underlying asset that is managed by the strategy on behalf of the specified address. It is used to track the assets held in the strategy.

Parameters

env: The execution environment.

from: The address for which to retrieve the balance (usually the DeFindex vault).

Returns

Ok(i128) containing the balance of the underlying asset.

Err(StrategyError) if the balance cannot be retrieved.

pub fn withdraw(&self, amount: &i128, from: &Address) -> i128

Withdraws the specified amount of the underlying asset from the strategy.

This function allows the DeFindex vault to withdraw assets from the strategy.

Parameters

env: The execution environment.

amount: The amount of the underlying asset to withdraw.

from: The address from which to withdraw the asset (usually the DeFindex vault).

Returns

Ok(i128) containing the amount withdrawn.

Err(StrategyError) if the withdrawal fails (e.g., insufficient balance).

pub fn try_withdraw( &self, amount: &i128, from: &Address ) -> Result<Result<i128, <i128 as TryFromVal<Env, Val>>::Error>, Result<StrategyError, InvokeError>>

Withdraws the specified amount of the underlying asset from the strategy.

This function allows the DeFindex vault to withdraw assets from the strategy.

Parameters

env: The execution environment.

amount: The amount of the underlying asset to withdraw.

from: The address from which to withdraw the asset (usually the DeFindex vault).

Returns

Ok(i128) containing the amount withdrawn.

Err(StrategyError) if the withdrawal fails (e.g., insufficient balance).