Struct defindex_strategy_core::Spec

pub struct Spec;

Implementations

impl Spec

pub const fn spec_xdr_initialize() -> [u8; 752]

Initializes the strategy with the given parameters.

This function is called to initialize the strategy with the required parameters. It sets up the strategy’s underlying asset and any additional parameters required for the strategy.

Parameters

env: The execution environment.

asset: The address of the underlying asset that this strategy will manage.

init_args: A vector of additional arguments, which may vary depending on the strategy’s requirements (e.g., configurations, addresses).

Returns

Ok(()) if the strategy is successfully initialized.

Err(StrategyError) if an error occurs during initialization.

impl Spec

pub const fn spec_xdr_asset() -> [u8; 468]

Returns the underlying asset address managed by the strategy.

This function provides the address of the asset that this strategy is managing. It is used by the DeFindex vault to determine which asset is associated with this strategy.

Parameters

env: The execution environment.

Returns

Ok(Address) containing the asset address.

Err(StrategyError) if the asset address cannot be retrieved.

impl Spec

pub const fn spec_xdr_deposit() -> [u8; 572]

Deposits the specified amount of the asset into the strategy.

This function allows the DeFindex vault to deposit assets into the strategy.

Parameters

env: The execution environment.

amount: The amount of the underlying asset to deposit into the strategy.

from: The address from which the asset will be transferred (usually the DeFindex vault).

Returns

Ok(()) if the deposit is successful.

Err(StrategyError) if the deposit fails for any reason.

impl Spec

pub const fn spec_xdr_harvest() -> [u8; 584]

Harvests yields generated by the strategy.

This function triggers any necessary actions required to generate or claim yields from the underlying DeFi protocol or strategy. It can be used to reinvest rewards, claim interest, or perform other actions that maximize returns.

Parameters

env: The execution environment.

from: The address that is triggering the harvest (usually the DeFindex vault).

Returns

Ok(()) if the harvest is successful.

Err(StrategyError) if the harvest fails.

impl Spec

pub const fn spec_xdr_balance() -> [u8; 600]

Returns the current balance of the underlying asset in the strategy.

This function provides the current balance of the underlying asset that is managed by the strategy on behalf of the specified address. It is used to track the assets held in the strategy.

Parameters

env: The execution environment.

from: The address for which to retrieve the balance (usually the DeFindex vault).

Returns

Ok(i128) containing the balance of the underlying asset.

Err(StrategyError) if the balance cannot be retrieved.

impl Spec

pub const fn spec_xdr_withdraw() -> [u8; 580]

Withdraws the specified amount of the underlying asset from the strategy.

This function allows the DeFindex vault to withdraw assets from the strategy.

Parameters

env: The execution environment.

amount: The amount of the underlying asset to withdraw.

from: The address from which to withdraw the asset (usually the DeFindex vault).

Returns

Ok(i128) containing the amount withdrawn.

Err(StrategyError) if the withdrawal fails (e.g., insufficient balance).